[LinuxConsumption] Support mounting Azure file share with Run-From-Package
Currently either Azure File share can be mounted or app can use Run-From-Package but not both simultaneously. This PR adds support for this scenario but only for PowerShell apps.
If AzureFiles are configured, it will be mounted at /home. If Run-From-Pkg is configured the contents will be made available at /home/site/wwwroot.
Note: It is possible for /home/site/wwwroot to point to a shared remote share. So copying contents to /home/site/wwwroot can be disruptive for other running instances. As a workaround /local/sitepackages will be used as a per-instance local storage where the zip will extracted before it gets mounted to /home/site/wwwwroot.

Fixes https://github.com/Azure/azure-functions-host/issues/7252