Defer Script Host disposal during cold start
One of the steps during specialization is to restart host. This restart involves starting a new host and disposing the old host.
https://github.com/Azure/azure-functions-host/blob/dev/src/WebJobs.Script.WebHost/WebJobsScriptHostService.cs#L483

Insight 1 - From profiling, disposal of the old host is the expensive part in the restart operation. Currently this is being run as part of a fire and forget task and there are no guarantees on when this task actually completes. meaning we have the flexibility to enforce some ordering here.
https://github.com/Azure/azure-functions-host/blob/dev/src/WebJobs.Script.WebHost/WebJobsScriptHostService.cs#L686

Insight 2 - On a single core machine, while the restart is in progress (including the expensive part mentioned above), the actual execution of function invocation is getting blocked at higher percentiles of cold start. On most cold starts the expensive parts of restart seems to be getting scheduled only after the end-end cold start has already completed. So the impact is only seen on relatively smaller number of cold starts but the impact is very high if the ordering was reversed .

Fix is to defer the Host disposal by 5 seconds. The fix attempts to defer the minimal code that provides biggest gains to reduce the surface area of the change.

Results:

Without the fix:
P10 - 735
P50 - 938
P90 - 2953

With the fix:
P10 - 735
P50 - 906
P90 - 1125

As noted above the biggest gains are seen at higher percentiles.